refactor into a plotting_tools file
